Relocate COVIDTestList
